Tweaks after self review
